wasmtime::component::Func: add a `ty(&self, store) -> ComponentFunc` getter
analog of wasmtime::Func::ty, this gets a representation of the function's type
which itself doesnt require a &mut store passed to the params/results getters.

by rewriting Func::params and Func::results in terms of this new common
getter, I could eliminate the Arc<ComponentTypes> element in the Func struct.